Перенёс урлы в отдельный класс
