+++ data/auth_data.py
@@ -1,10 +1,4 @@
-class TestUrlData:
-    URL = 'https://stellarburgers.nomoreparties.site/'
-    URL_LOG = 'login'
-    URL_FEED = 'feed'
-    URL_PROFILE = 'account/profile'
-    URL_FORGOT_PASSWORD = 'forgot-password'
-    URL_ORDER_HISTORY = 'account/order-history'
+class TestAuthData:
     COR_NAME = 'Zemdan'
     COR_EMAIL = '1234@ya.com'
     COR_PASSWORD = '123ddd'
