+++ data/url_data.py
@@ -1,0 +1,7 @@
+class TestUrlData:
+    URL = 'https://stellarburgers.nomoreparties.site/'
+    URL_LOG = 'login'
+    URL_FEED = 'feed'
+    URL_PROFILE = 'account/profile'
+    URL_FORGOT_PASSWORD = 'forgot-password'
+    URL_ORDER_HISTORY = 'account/order-history'
